Bugfixes
Too large documentsize was throwing off latex typesetting so I made it a bit smaller and adjustable from the commandline.

There was also a bug where the parser would stop parsing if it thought the program was done and the input wasn't exhausted. In reality this often means a semicolon was forgotten, and it caused unintended behaviour so now it just throws an exception.