Add --init-state flag
Added a flag that allows for an initialisation of the state. Usefull for some assignments, like exercise 7 from week 1 where we had to derive a tree assuming the starting state had values 23 and 6 for x and y respectively.

If you would prepend x:=23; y :=6; ...
to the program, the derivation tree would grow which we do not want, so this flag is necessary.